requirements.txt
webbrowser
yaml
flask
requests
selenium
bs4
pillow
opencv-python
webdriver_manager